+++ eras/alonzo/impl/testlib/Test/Cardano/Ledger/Alonzo/Arbitrary.hs
@@ -66,13 +66,7 @@ import Cardano.Ledger.Alonzo.TxWits (
   TxDats (TxDats),
  )
 import Cardano.Ledger.BaseTypes (StrictMaybe (..))
-import Cardano.Ledger.Plutus.Data (
-  BinaryData,
-  Data (..),
-  Datum (..),
-  dataToBinaryData,
-  hashData,
- )
+import Cardano.Ledger.Plutus.Data (hashData)
 import Cardano.Ledger.Plutus.ExUnits (ExUnits (..))
 import Cardano.Ledger.Plutus.Language (
   Language (..),
@@ -101,12 +95,6 @@ import Test.Cardano.Ledger.Core.Arbitrary (
  )
 import Test.Cardano.Ledger.Mary.Arbitrary ()
 import Test.Cardano.Ledger.Plutus (alwaysFailsPlutus, alwaysSucceedsPlutus)
-
-instance Era era => Arbitrary (Data era) where
-  arbitrary = Data <$> arbitrary
-
-instance Era era => Arbitrary (BinaryData era) where
-  arbitrary = dataToBinaryData <$> arbitrary
 
 instance Arbitrary PV1.Data where
   arbitrary = resize 5 (sized gendata)
@@ -444,17 +432,6 @@ instance
       <*> arbitrary
       -- FIXME: why singleton? We should generate empty as well as many value sets
       <*> (Set.singleton <$> (getLanguageView @era <$> arbitrary <*> arbitrary))
-
-instance
-  Era era =>
-  Arbitrary (Datum era)
-  where
-  arbitrary =
-    oneof
-      [ pure NoDatum
-      , DatumHash <$> arbitrary
-      , Datum . dataToBinaryData <$> arbitrary
-      ]
 
 deriving instance Arbitrary CoinPerWord